requirements.txt
aiohttp==3.11.11
httpx==0.28.1
matplotlib==3.10.0
numpy==2.2.0
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pydeck==0.9.1
requests==2.32.3
scikit-learn==1.6.0
seaborn==0.13.2
streamlit==1.41.1